Remove vilpy-follow (gd)
Remove vilpy-follow (`F` in lispy, `gd` in vilpy)

This function used to go to the defintion of the symbol after the parenthesis.

Doing so has a couple of issue, and for now, it's safer simply not to implement this fucntionaly in `vilpy`.

== It's out of scope
. `vilpy` scope should be restricted to nagivation and code transformation anyway. So going to definition is out of scope.

== It's hard to know what to expect
. It is hard to guess what to expect from this command. `xref-find-definitions` looks reasonable, but there not all modes implement it yet. For example, if I'm not mistaken, `cider` does not use it.
One could argue: Why not use `cider` owns command if `cider-mode` is active? The issue is that even if its mode is active, you might want to rely on LSP instead. In fact, this is my personal expectation.

To complicate things further, people that use Doom, Spacemacs or other Emacs distributions might want to use their own mechanisms.

== `xref-find-definitions` is difficult to get right in different modes
Let's suppose that all relevant major modes implement `xref-find-definitions`.
Still, getting it right in `vilpy` is not easy, because it expects the symbol to be at point, and calling it programatically might have unexpected results.
For example, consider a Clojure buffer in a project that is running `eglot` (LSP client). Say the point is just before the parenthesis, in `vilpy special`. Then, calling `xref-find-definitions` with the next symbol won't work, because `eglot` sends the point position to the LSP server and it gets it wrong. (we could move the point forward and use `xref-find-definitions`. It works, but (a) the elisp would ask for a symbol and (b) the point position to be restored with `C-,` would be after the parenthesis, which is awkward for `vilpy`).

== Conclusion
Maybe this feature can be readded in the future, especially with further investigation from `xref-find-definitions`. For now, we can live without it.
For going to definition of a symbol when the point is just before parenthesis, simply use `C-f C-.`.